Source: https://github.com/msupply-foundation/open-msupply/commit/c8fd5830d698a4c5444202045057293a1af477bc

+++ server/repository/src/db_diesel/pack_variant.rs
@@ -23,7 +23,7 @@ pub struct PackVariantRepository<'a> {
 pub struct PackVariantFilter {
     pub id: Option<EqualFilter<String>>,
     pub item_id: Option<EqualFilter<String>>,
-    pub pack_size: Option<EqualFilter<i32>>,
+    pub pack_size: Option<EqualFilter<f64>>,
     pub is_active: Option<bool>,
 }
 
@@ -112,7 +112,7 @@ impl PackVariantFilter {
         self
     }
 
-    pub fn pack_size(mut self, filter: EqualFilter<i32>) -> Self {
+    pub fn pack_size(mut self, filter: EqualFilter<f64>) -> Self {
         self.pack_size = Some(filter);
         self
     }
